\name{ROI.plugin.deoptimr-package}
\docType{package}
\alias{ROI.plugin.deoptimr-package}

\title{deoptimr}

\description{
This package is part of the \href{https://cran.r-project.org/package=ROI}{R Optimization Infrastructure}
\pkg{ROI}
}

\references{
  Babu, B. V. and Angira, R. (2006)
  Modified differential evolution (MDE) for optimization of
  non-linear chemical processes.
  \emph{Computers and Chemical Engineering} \bold{30}, 989--1002.

  Brest, J., Greiner, S., Boskovic, B., Mernik, M. and Zumer, V. (2006)
  Self-adapting control parameters in differential evolution: a
  comparative study on numerical benchmark problems.
  \emph{IEEE Transactions on Evolutionary Computation} \bold{10}, 646--657.

  Lampinen, J. and Zelinka, I. (1999).
  Mechanical engineering design optimization by differential evolution;
  in Corne, D., Dorigo, M. and Glover, F., Eds.,
  \emph{New Ideas in Optimization}.
  McGraw-Hill, pp. 127--146.

  Price, K. V., Storn, R. M. and Lampinen, J. A. (2005)
  \emph{Differential Evolution: A practical approach to global optimization}.
  Springer, Berlin, pp. 117--118.

  Storn, R. (2008)
  Differential evolution research --- trends and open questions;
  in Chakraborty, U. K., Ed.,
  \emph{Advances in differential evolution}.
  SCI 143, Springer-Verlag, Berlin, pp. 11--12.

  Storn, R. and Price, K. (1997)
  Differential evolution - a simple and efficient heuristic for global
  optimization over continuous spaces.
  \emph{Journal of Global Optimization} \bold{11}, 341--359.

  Wu, G., Pedrycz, W., Suganthan, P. N. and Mallipeddi, R. (2015)
  A variable reduction strategy for evolutionary algorithms handling
  equality constraints.
  \emph{Applied Soft Computing} \bold{37}, 774--786.

  Zhang, H. and Rangaiah, G. P. (2012)
  An efficient constraint handling method with integrated differential
  evolution for numerical and engineering optimization.
  \emph{Computers and Chemical Engineering} \bold{37}, 74--88.

  Zielinski, K. and Laur, R. (2008)
  Stopping criteria for differential evolution in constrained
  single-objective optimization;
  in Chakraborty, U. K., Ed.,
  \emph{Advances in differential evolution}.
  SCI 143, Springer-Verlag, Berlin, pp. 111--138.
}

\seealso{
  Function \href{https://cran.r-project.org/package=DEoptimR/DEoptimR.pdf}{\code{JDEoptim()}} in the
  \pkg{DEoptimR} package.
}
